requirements.txt
Flask
Pillow
requests
keras
numpy
tensorflow